Install scripts to a 'site_bin' directory.
I have a big problem here: pkg does not allow packages to install
conflicting files, and perl has no 'site' or 'vendor' bin directories.
This means that when I build a 'site' version of a module which exists
in the core and includes scripts, like Module::Build, the package
created cannot be installed.

For now I'm working around this with an explicit, hardcoded site_bin
directory. Obviously this needs changing at some point, but I want to
get other things working first.

I'm not sure what the right answer is. Obviously I need to detect this,
probably with my own list of files in each package, but I'm not sure
what to do about it. I can't disallow it, since more-recent-than-core
versions of Module::Build seem to be quite a common requirement. The
simplest solution from my point of view would be to remove the offending
files from the 'new' package; this is exactly the opposite of what a
standard install would do, and since the new modules will come first in
@INC will probably make the script in question break.