Include package's immediate deps.
This is not enough, at least not for pkg: packages should contain their
whole recursive list of deps, not just the immediate ones. Although
packages created like this install OK, the registered metadata is not
correct, so for instance the reverse-deps of an indirect dep will not
include this package.