Post-install scripts.
This is a gross hack, but I don't think I have any alternative until
CPAN acquires some relevant metadata. For EUMM distributions only, we
parse the Makefile and attempt to extract the commands that will run
post-install. This will always include the perllocal entry, and may
include other things. Specifically, the SAX modules all have a target to
register themselves with the SAX infrastructure.

The Makefile parsing is done with Makefile::Parser, which is a pretty
shoddy Makefile parser, but is the best I could find. About the only
important thing it seems to do wrong with EUMM Makefiles is expand
variables while reading, rather than recursively; this requires a bit
more hacking to work around.