Don't take name/version from the Dist.
Dists still have names, for convenience (for messages, and so we've got
something to call the wrkdir), but that name is no longer used to name
the created package. Instead, Build now has ->name and ->version, which
get the name and version out of META.yml.

Some of the internal methods (pkg_for_dist and install_my_pkgs) are now
a bit untidy. I think, probably, I need a Pkg class, to hold name/
version/origin information, to replace my old use of Dist for that
purpose.