*O*K, let's try this again...
Keep separate lists of modules tried and dists tried. If we've
previously tried either a module or a dist, then there's no point trying
it again. If we get a Needed exn with a module in that's already been
tried, then this is either a dep on a module that's already failed or a
circular dep; in either case, this build fails.

Keep a list of failed dists just to print at the end.