Extract more information from the core perl pkg.
This removes all instances of the code being specific to my setup.
There's still a bit of a hack in the case of site_bin being equal to bin
in the core Config: I can't see any way to resolve the file conflict, so
for now we make up our own site_bin directory.